Debug: tejuinput func input

+++ src/omv/modules/py_tf.c
@@ -432,7 +432,7 @@ STATIC mp_obj_t py_tf_tejuinput(uint n_args, const mp_obj_t *args, mp_map_t *kw_
     mp_obj_list_t *arg_list = args[1];
     for(size_t i =0; i < arg_list->len; i++){
         arr[i] = (float) mp_obj_float_get(arg_list->items[i]);
-        printf("%f \n", arr[i]);
+        printf("%f \n", (double) arr[i]);
     }
 
     printf("input arr ready\n");
